Pass installation notification not as payload to redirect
Previously, we passed a notification upon installing the agent
as a payload (parameter) to an redirect. This notification object
was basically just a HTML container with a message in it. The HTML
was urlencoded and appended to the GET "redirect" request which pointed
to the Dashboard. Apparently, Windows interprets the urlencoded HTML
differently than Linux albeit being the same string. Therefore, change
the way notifications are shown by using Plesk's addInfo interface
instead of redirector.